gateway/requirements.txt
fastapi
uvicorn[standard]
httpx
aiofiles
python-multipart
sse-starlette
pydantic-settings
pyyaml
